# backend/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Student Academic Management System API",
    "version": "1.0.0",
    "description": "API documentation for the AMS backend across Academic Master, Examination, Results & Backlogs, and Academic Support modules."
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Local Development Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter your JWT token in the format: Bearer <token>"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/v1/results-backlogs/risk-thresholds": {
      "post": {
        "summary": "Create Risk Threshold",
        "tags": [
          "RiskThreshold"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "level": {
                    "type": "string"
                  },
                  "condition": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          },
          "201": {
            "description": "Created successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden (Requires roles: )"
          },
          "404": {
            "description": "Not found"
          }
        }
      },
      "get": {
        "summary": "Get all Risk Thresholds",
        "tags": [
          "RiskThreshold"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "201": {
            "description": "Created successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden (Requires roles: )"
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
    "/api/v1/results-backlogs/risk-thresholds/{id}": {
      "get": {
        "summary": "Get Risk Threshold by ID",
        "tags": [
          "RiskThreshold"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "201": {
            "description": "Created successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden (Requires roles: )"
          },
          "404": {
            "description": "Not found"
          }
        }
      },
      "put": {
        "summary": "Update Risk Threshold by ID",
        "tags": [
          "RiskThreshold"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "condition": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          },
          "201": {
            "description": "Created successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden (Requires roles: )"
          },
          "404": {
            "description": "Not found"
          }
        }
      },
      "delete": {
        "summary": "Delete Risk Threshold by ID",
        "tags": [
          "RiskThreshold"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "201": {
            "description": "Created successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden (Requires roles: )"
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
    "/api/v1/results-backlogs/risk": {
      "get": {
        "summary": "Get all Risk Profiles",
        "tags": [
          "Risk"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "201": {
            "description": "Created successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden (Requires roles: )"
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
...
    "/api/v1/analytics/risk": {
      "get": {
        "summary": "Get Risk Distribution",
        "tags": [
          "Analytics"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "201": {
            "description": "Created successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden (Requires roles: )"
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
...
  },
...
}